Rename & clean adapter.crypto and adapter.shared
