+++ packages/driver/src/cryptoUtils.ts
