+++ packages/driver/src/index.shared.ts
@@ -29,7 +29,7 @@ export {
 export { ConfigMemory } from "./datatypes/memory";
 export { Range, MultiRange } from "./datatypes/range";
 export { SparseVector } from "./datatypes/pgvector";
-export { Float16Array } from "./adapter.shared";
+export { Float16Array } from "./utils";
 
 export type { Executor } from "./ifaces";
