+++ packages/driver/test/scram.test.ts
@@ -19,7 +19,7 @@
 import { getSCRAM, saslprep } from "../src/scram";
 import { cryptoUtils as nodeCryptoUtils } from "../src/nodeCrypto";
 import { cryptoUtils as browserCryptoUtils } from "../src/browserCrypto";
-import cryptoUtils from "../src/adapter.crypto";
+import cryptoUtils from "../src/cryptoUtils";
 
 async function generateScramWith(scram: ReturnType<typeof getSCRAM>) {
   // Test SCRAM-SHA-256 against an example in RFC 7677
